pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "VictoorLDS", email = "vlds221005@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
    "django-cors-headers>=4.2.0",
    "django-extensions>=3.2.3",
    "pydotplus>=2.0.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}
